corrigi uns exercicios

+++ Aula11/desafio40.py
@@ -1,5 +1,5 @@
-nota1 = int(input('Primeira nota: '))
-nota2 = int(input('Segunda nota: '))
+nota1 = float(input('Primeira nota: '))
+nota2 = float(input('Segunda nota: '))
 media = (nota1 + nota2) / 2
 print('Tirando {} e {}, a média do aluno é {}'.format(nota1, nota2, media))
 if media < 5:
